Move hazard_info_popup to map_page (misunderstood its purpose last time) Rename hazard_info_card back to hazard_info (also misunderstood its purpose)
